colocated types
